// frontend/client/src/features/profile/info.ts
export interface UserInfo {
    email: string | undefined;
    username: string | undefined;
    name: string | undefined;
    surname: string | undefined;
    city: string | undefined;
    street: string | undefined;
    number: number | undefined;
    phoneNumber: number | undefined;
    additionalInfo: string | undefined;
}